refactor(core): promote pkg/authcheck to internal/core/authcheck (no behavior change)
