+++ cmd/databricks-claude/serve.go
@@ -14,7 +14,7 @@ import (
 	"time"
 
 	"github.com/IceRhymers/databricks-agents/internal/cmd"
-	"github.com/IceRhymers/databricks-agents/pkg/authcheck"
+	"github.com/IceRhymers/databricks-agents/internal/core/authcheck"
 	"github.com/IceRhymers/databricks-agents/pkg/mdmprofile"
 	"github.com/IceRhymers/databricks-agents/pkg/proxy"
 	"github.com/IceRhymers/databricks-agents/pkg/websearch"
