+++ cmd/databricks-claude/serve_session.go
@@ -11,9 +11,9 @@ import (
 	"os/signal"
 	"syscall"
 
+	"github.com/IceRhymers/databricks-agents/internal/core/authcheck"
 	"github.com/IceRhymers/databricks-agents/internal/core/portbind"
 	"github.com/IceRhymers/databricks-agents/internal/core/refcount"
-	"github.com/IceRhymers/databricks-agents/pkg/authcheck"
 	"github.com/IceRhymers/databricks-agents/pkg/health"
 	"github.com/IceRhymers/databricks-agents/pkg/lifecycle"
 	"github.com/IceRhymers/databricks-agents/pkg/proxy"
