+++ internal/core/authcheck/AGENTS.md
